Implement searching alias field
Added the code necessary to also do a simultaneous regex search on
field 11 (technically "10646 comment", but usually just an alias for
the character name).

This is needed because some common names are only shown as aliases.
For example, "backslash" is known in unicode as "REVERSE SOLIDUS".

The regex necessary to do this was big and hairy and made me
appreciate AWK more. (It's trivial to do this in awk).