add support for ^ and $ anchors
I took for granted that awk let me use ^ and $ to search for
the beginning and ending of fields instead of lines. I tried to
reimplement that in Python, but it's not quite right as it only
checks the first and last character. For example, ugrep "^x" works,
but ugrep "(^x)" does not.

This is getting uglier and I'm beginning to think I'm going to have
make a loop to search this file line by line. Trés annoying.