Extract category from UnicodeData.txt and determine isprint()
The default isprintable builtin in Python can be out-of-date. So,
instead we parse the third field, "General Category". Ugrep's new
isprint() returns True unless the caegory begins with 'C'.

   Cc   Control               a C0 or C1 control code
   Cf   Format                a format control character
   Cs   Surrogate             a surrogate code point
   Co   Private_Use           a private-use character
   Cn   Unassigned            a reserved unassigned code point or a noncharacter
   C    Other                 Cc | Cf | Cs | Co | Cn

Also added iscombining(). If the category begins with 'M' it is a
combining character and we should print a space first for the
character to combine with.

  Mn   Nonspacing_Mark       a nonspacing combining mark (zero advance width)
  Mc   Spacing_Mark          a spacing combining mark (positive advance width)
  Me   Enclosing_Mark        an enclosing combining mark
  M    Mark                  Mn | Mc | Me