Handle Python3's silly Exception ignored BrokenPipeError
Even if you catch a broken pipe error on stdout, when Python exits, it
always tries to flush stdout which throws another error which barfs
out this:

Exception ignored in: <_io.TextIOWrapper name='<stdout>' mode='w' encoding='utf-8'>
BrokenPipeError: [Errno 32] Broken pipe

This is silly because we already knew that stdout couldn't be flushed
because we got a broken pipe. We can work around it by setting
stdout = None.