Possibly interpret range {10..8} differently than 10..8
Normally 10..8 means 10 11 12 13 14 15 16 17 18

This makes a lot of sense for ugrep as you usually don't want to have
to type the whole first hexadecimal number again. E.g., U+2100..8.

However, bash's brace expansion doesn't allow the shortcut of
specifying just the last few digits. So, in this commit, I've made
ugrep detect if braces are being used and then handle them the same
way.

I'll probably undo this, but I'm adding it to the git log for later.