Add routine to check if character should be in a box
The Unicode charts put some characters in dashed boxes. There are some
rules, but it is also somewhat arbitrary. I will probably use this
mark white space characters of various widths.

This patch also leaves the fc-list sort unchanged so the names should
be in the order the systam actually picks the fonts.