Cache rows/cols, print 2nd nl when convert dies on NUL with DECDHL
Rows/cols are requested every time. To cut down on ioctls, cache the
result. Later we can add a SIGWINCH handler if it needs to change as
the script is run.

NULL character cannot be sent to convert using command line. Normally
we just send a single new line, but we should send two if our font is
double high.