A simple regression test
Since ugrep will be rewritten someday now that it is clearer what it
should do, it makes sense to have a regression test to detect if any
bugs are introduced.

This script is very simple: it just runs commands and checks if the
number of lines output is equal to number expected.