Ensure each regex matches exactly one line
Python's regex's allow '.' to match newline. That meant if the user
passed in something like:

     ugrep '.*'

Then the <codepoint> portion of the regex could match semicolons and
would slurp up the rest of the data: '0000;<control>;Cc;0;BN;;;;;N;NULL;;;'.

The solution is to force a strict accounting of the semicolons as we
know exactly how many there will be on every line.

For belt-and-suspenders, anywhere '[^;]*' was used to mean "anything
but a semicolon", has been changed to '[^;\n]*'